The map now opens centered on the user’s location
If the user is off campus, they can simply click the new “Paw” button, and it moves the map’s camera to the Student Union on campus

The user now has the ability to get directions to a location. They simply have to click on the location and will be given the option to switch to GMaps for directions.

Fixed a bug where the circular ProgressBar wasn’t updating unless you close and reopen the app.